refactor error handling and processing the list of keys
